+++ omnibenchmark/cli/archive.py
@@ -71,8 +71,8 @@ from .debug import add_debug_option
     show_default=True,
 )
 @click.option(
-    "-rs",
-    "--remote-storage",
+    "-r",
+    "--use-remote-storage",
     help="Execute and store results remotely. Default False.",
     is_flag=True,
     default=False,
@@ -90,16 +90,14 @@ def archive(
     compression,
     compresslevel,
     dry_run,
-    remote_storage,
+    use_remote_storage,
     out_dir,
 ):
     """Archive a benchmark and its artifacts."""
 
     # Validate out_dir usage
-    if remote_storage and out_dir != "out":
-        logger.error(
-            "-Invalid arguments: --out-dir can only be used without --remote-storage"
-        )
+    if use_remote_storage and out_dir != "out":
+        logger.error("Invalid arguments: --out-dir can only be used with local storage")
         sys.exit(1)
 
     assert benchmark is not None
@@ -129,7 +127,7 @@ def archive(
         compression=compression,
         compresslevel=compresslevel,
         dry_run=dry_run,
-        remote_storage=remote_storage,
+        remote_storage=use_remote_storage,
     )
     if dry_run:
         click.echo(f"Files to archive:\n{tree_string_from_list(archive_file)}")