+++ omnibenchmark/cli/run.py
@@ -75,8 +75,8 @@ def run(ctx):
     default=False,
 )
 @click.option(
-    "-rs",
-    "--remote-storage",
+    "-r",
+    "--use-remote-storage",
     help="Execute and store results remotely. Default False.",
     is_flag=True,
     default=False,
@@ -109,7 +109,7 @@ def run_benchmark(
     update,
     dry,
     yes,
-    remote_storage,
+    use_remote_storage,
     keep_module_logs,
     continue_on_error,
     task_timeout,
@@ -133,10 +133,8 @@ def run_benchmark(
             sys.exit(1)
 
     # Validate out_dir usage
-    if remote_storage and out_dir != "out":
-        logger.error(
-            "-Invalid arguments: --out-dir can only be used without --remote-storage"
-        )
+    if use_remote_storage and out_dir != "out":
+        logger.error("Invalid arguments: --out-dir can only be used with local storage")
         sys.exit(1)
 
     try:
@@ -161,7 +159,7 @@ def run_benchmark(
         msg = "re-run the entire workflow"
         abort_if_user_does_not_confirm(msg, logger)
 
-    if remote_storage:
+    if use_remote_storage:
         storage_options = remote_storage_snakemake_args(b)
         # creates bucket if it doesn't exist
         _ = get_storage_from_benchmark(b)